Move the message encoding/decoding and other stuff to a separate file
